(feat): create vim layer
- a basic vim layer is created. Using this layer, I will emulate basic
vim functionality (normal mode) for all programs, instead of just in vim
by using appropriate keys. Another vim insert layer is necessary, which
should be not much else than the BASE layer, except that ESC is replaced
with a layer switch to the VIM layer. To preserve full functionality in
vim, there is an easy way to go to the BASE layer, which can then be
used normally in vim.